Update before copy of backend

+++ Models/Contract/IViewModel.cs
@@ -1,8 +1,8 @@
 ﻿namespace Models.Contract
 {
     public interface IViewModel
-    {        
-        int UserId { get; set; }
+    {
+        long UserId { get; set; }
         IFormEntity ConvertToEntity();
     }
 }
